requirements.txt
aiohttp==3.7.4.post0
aiosqlite==0.17.0
aiotrivia==2.0.1
akinator.py==5.0.0
animec==0.4
async-timeout==3.0.1
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.10.0
bella==2.1.5
boto==2.49.0
bs4==0.0.1
cargo==0.3
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
cogwatch==2.1.0
colorama==0.4.4
dataclasses-json==0.5.6
discord-components==2.1.2
discord-ext-menus==1.1
discord-pretty-help==1.3.3
discord.py==1.7.3
discordbotdash==1.2.3
enum34==1.1.10
fire==0.3.1
Flask==2.0.3
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
isodate==0.6.1
itsdangerous==2.1.0
Jinja2==3.0.3
ksoftapi==0.4.1
MarkupSafe==2.1.0
marshmallow==3.14.1
marshmallow-enum==1.5.1
multidict==6.0.2
mypy-extensions==0.4.3
mysql==0.0.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
oauthlib==3.2.0
Pillow==9.0.1
praw==7.5.0
prawcore==2.3.0
protobuf==3.19.4
pycodestyle==2.8.0
pycparser==2.21
PyNaCl==1.4.0
python-dotenv==0.19.2
python-youtube==0.6.4
pytube3==9.6.4
pytz==2021.3
randfacts==0.15.0
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
sr-api==1.1.3
termcolor==1.1.0
toml==0.10.2
typing-inspect==0.7.1
typing_extensions==4.1.1
ujson==4.3.0
update-checker==0.18.0
urllib3==1.26.8
watchgod==0.7
websocket-client==1.2.3
Werkzeug==2.0.3
yarl==1.7.2
youtube-dl==2021.12.17
ytdl==1.0.0rc5
